+++ instrumentation/spring/spring-scheduling-3.1/javaagent/src/test/java/io/opentelemetry/javaagent/instrumentation/spring/scheduling/v3_1/spring/component/IntervalTask.java
@@ -3,7 +3,7 @@
  * SPDX-License-Identifier: Apache-2.0
  */
 
-package spring.component;
+package io.opentelemetry.javaagent.instrumentation.spring.scheduling.v3_1.spring.component;
 
 import java.util.concurrent.CountDownLatch;
 import java.util.concurrent.TimeUnit;
@@ -11,11 +11,11 @@ import org.springframework.scheduling.annotation.Scheduled;
 import org.springframework.stereotype.Component;
 
 @Component
-public class TriggerTask implements Runnable {
+public class IntervalTask implements Runnable {
 
   private final CountDownLatch latch = new CountDownLatch(1);
 
-  @Scheduled(cron = "0/5 * * * * *")
+  @Scheduled(fixedRate = 5000)
   @Override
   public void run() {
     latch.countDown();
