+++ instrumentation/spring/spring-scheduling-3.1/javaagent/src/test/java/io/opentelemetry/javaagent/instrumentation/spring/scheduling/v3_1/spring/component/OneTimeTask.java
@@ -3,19 +3,17 @@
  * SPDX-License-Identifier: Apache-2.0
  */
 
-package spring.component;
+package io.opentelemetry.javaagent.instrumentation.spring.scheduling.v3_1.spring.component;
 
 import java.util.concurrent.CountDownLatch;
 import java.util.concurrent.TimeUnit;
-import org.springframework.scheduling.annotation.Scheduled;
 import org.springframework.stereotype.Component;
 
 @Component
-public class IntervalTask implements Runnable {
+public class OneTimeTask implements Runnable {
 
   private final CountDownLatch latch = new CountDownLatch(1);
 
-  @Scheduled(fixedRate = 5000)
   @Override
   public void run() {
     latch.countDown();
