+++ instrumentation/spring/spring-scheduling-3.1/javaagent/src/test/java/io/opentelemetry/javaagent/instrumentation/spring/scheduling/v3_1/spring/component/TaskWithError.java
@@ -3,7 +3,7 @@
  * SPDX-License-Identifier: Apache-2.0
  */
 
-package spring.component;
+package io.opentelemetry.javaagent.instrumentation.spring.scheduling.v3_1.spring.component;
 
 import java.util.concurrent.CountDownLatch;
 import java.util.concurrent.TimeUnit;
