+++ instrumentation/spring/spring-scheduling-3.1/javaagent/src/test/java/io/opentelemetry/javaagent/instrumentation/spring/scheduling/v3_1/spring/config/TriggerTaskConfig.java
@@ -3,18 +3,18 @@
  * SPDX-License-Identifier: Apache-2.0
  */
 
-package spring.config;
+package io.opentelemetry.javaagent.instrumentation.spring.scheduling.v3_1.spring.config;
 
+import io.opentelemetry.javaagent.instrumentation.spring.scheduling.v3_1.spring.component.TriggerTask;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
 import org.springframework.scheduling.annotation.EnableScheduling;
-import spring.component.IntervalTask;
 
 @Configuration
 @EnableScheduling
-public class IntervalTaskConfig {
+public class TriggerTaskConfig {
   @Bean
-  public IntervalTask scheduledTasks() {
-    return new IntervalTask();
+  public TriggerTask triggerTasks() {
+    return new TriggerTask();
   }
 }
